refactor(client): move utils module to improve organization and add header partitioning function
