+++ u-sdk/src/oss/object/mod.rs
@@ -1,6 +1,5 @@
 mod basic;
 mod multipart_upload;
 mod types_rs;
-pub(crate) mod utils;
 
 pub use types_rs::*;
